Fix coverity issues 382921 382924 382927 382928 382932 382933 382950 382990 383123 382952 382906 382908 382912 382914 382917 382918 382919

+++ database/engine/cache.c
@@ -1946,7 +1946,7 @@ time_t pgc_page_update_every_s(PGC_PAGE *page) {
 
 time_t pgc_page_fix_update_every(PGC_PAGE *page, time_t update_every_s) {
     if(page->update_every_s == 0)
-        page->update_every_s = update_every_s;
+        page->update_every_s = (uint32_t) update_every_s;
 
     return page->update_every_s;
 }
